Move DFC patches into our DFC engine
It seems to belong there and it's avoiding some loading issues with
Spring.
